+++ src/components/dropdown/index.js
@@ -1,3 +1,7 @@
 import Dropdown from './Dropdown'
+import DropdownOption from './DropdownOption'
 
-export default Dropdown
+export {
+    Dropdown,
+    DropdownOption
+}