+++ src/components/select/index.js
@@ -1,7 +1,3 @@
 import Select from './Select'
-import Option from './Option'
 
-export {
-    Select,
-    Option
-}
+export default Select